fix: add new directive to overwrite existing templates
