+++ boilerplates/mantine/files/$README.md.ts
@@ -5,7 +5,17 @@ export default async function getReadme(props: TransformerProps) {
 
   //language=Markdown
   const todo = `
-## *Mantine*
+## Mantine
+
+This is a boilerplate for Mantine based on the [Getting Started](https://mantine.dev/docs/getting-started/) guide.
+
+The following Packages are installed:
+* \`@mantine/hooks\` Hooks for state and UI management
+* \`@mantine/core\` Core components library: inputs, buttons, overlays, etc.
+
+If you add more packages, make sure to update the \`layouts/LayoutDefault.tsx\` file to include the required CSSs.
+
+The theme is defined in \`layouts/theme.js\`.
 
 `;
