+++ boilerplates/mantine/package.json
@@ -17,10 +17,16 @@
     "@mantine/core": "^7.12.1",
     "@mantine/hooks": "^7.12.1",
     "@types/node": "^18.19.14",
+    "@types/react": "^18.3.4",
+    "@types/react-dom": "^18.3.0",
     "postcss": "^8.4.41",
     "postcss-preset-mantine": "^1.17.0",
     "postcss-simple-vars": "^7.0.1",
-    "react": "^18.3.1"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "vike-react": "^0.5.3",
+    "vite": "^5.4.2",
+    "vite-plugin-compiled-react": "^1.1.3"
   },
   "dependencies": {
     "@batijs/core": "workspace:*"
@@ -29,6 +35,7 @@
     "dist/"
   ],
   "bati": {
+    "enforce": "post",
     "if": {
       "flag": {
         "$all": [
@@ -42,8 +49,14 @@
     "./postcss.config.": {
       "types": "./dist/types/postcss.config.d.cts"
     },
+    "./components/Link": {
+      "types": "./dist/types/components/Link.d.ts"
+    },
     "./layouts/theme": {
       "types": "./dist/types/layouts/theme.d.ts"
+    },
+    "./layouts/LayoutDefault": {
+      "types": "./dist/types/layouts/LayoutDefault.d.ts"
     }
   },
   "typesVersions": {
@@ -51,8 +64,14 @@
       "postcss.config.": [
         "./dist/types/postcss.config.d.cts"
       ],
+      "components/Link": [
+        "./dist/types/components/Link.d.ts"
+      ],
       "layouts/theme": [
         "./dist/types/layouts/theme.d.ts"
+      ],
+      "layouts/LayoutDefault": [
+        "./dist/types/layouts/LayoutDefault.d.ts"
       ]
     }
   }
