+++ boilerplates/react/package.json
@@ -21,6 +21,7 @@
     "@batijs/telefunc": "workspace:^",
     "@batijs/trpc": "workspace:^",
     "@batijs/ts-rest": "workspace:*",
+    "@mantine/core": "^7.12.1",
     "@types/node": "^18.19.14",
     "@types/react": "^18.3.4",
     "@types/react-dom": "^18.3.0",
